timer

+++ src/Dashboard.jsx
@@ -17,6 +17,10 @@ const Dashboard = () => {
       }
       const decoded = jwtDecode(token);
       console.log(decoded);
+      const data = {
+        userId: 1,
+      }
+      getPackageQuestions(data);
     } catch (error) {
       console.log(error.message);
     }
@@ -24,8 +28,17 @@ const Dashboard = () => {
 
   const getPackageQuestions = async () => {
     try {
-      const response = await axios.get(`https://sbpmb-express.amisbudi.cloud/packagequestions`);
-      setPackages(response.data.data);
+      const responsePackageQuestions = await axios.get(`http://localhost:3000/packagequestions`);
+      const responseRecords = await axios.get(`http://localhost:3000/records`);
+
+      const packageQuestions = responsePackageQuestions.data.data;
+      const records = responseRecords.data;
+
+      const filteredPackageQuestions = packageQuestions.filter((pq) => {
+        return !records.some((record) => record.package_question_id === pq.id && record.user_id === 1);
+      });
+
+      setPackages(filteredPackageQuestions);
     } catch (error) {
       console.log(error.message);
     }
@@ -38,74 +51,73 @@ const Dashboard = () => {
     }
   }
 
-  const checkoutPackage = async (id) => {
+  const checkoutPackage = async (pkg) => {
     try {
-      const activePackage = localStorage.getItem('CBT:package');
-      if (activePackage) {
-        navigate('/assesment');
-      } else {
-        const data = {
-          package_question_id: id,
-          user_id: 1,
+      if (window.confirm(`Apakah anda yakin akan memulai tes ${pkg.name}?`)) {
+        const activePackage = localStorage.getItem('CBT:package');
+        if (activePackage) {
+          navigate('/assesment');
+        } else {
+          const data = {
+            package_question_id: pkg.id,
+            user_id: 1,
+          }
+          localStorage.setItem('CBT:package', JSON.stringify(data));
+          await axios.post(`http://localhost:3000/questionusers`, data);
+          navigate('/assesment');
         }
-        localStorage.setItem('CBT:package', JSON.stringify(data));
-        const response = await axios.post(`https://sbpmb-express.amisbudi.cloud/questionusers`, data);
-        navigate('/assesment');
       }
     } catch (error) {
       console.log(error.message);
     }
   }
 
+  const handleLogout = () => {
+    if (window.confirm('Apakah anda yakin akan keluar?')) {
+      const token = localStorage.getItem('CBTtrisakti:token');
+      if (!token) {
+        alert('Mohon maaf, token tidak valid!');
+      } else {
+        localStorage.removeItem('CBTtrisakti:token');
+        navigate('/')
+      }
+    }
+  }
+
   useEffect(() => {
     AOS.init({
-      duration: 500,
+      duration: 600,
       easing: "ease-in-out",
     });
     getInfo();
-    getPackageQuestions();
     checkAssesment();
   }, []);
   return (
-    <div className="bg-gradient-to-b from-[#005D99] to-[#005083] h-screen flex-col lg:flex-row lg:flex items-center justify-center p-12 text-white gap-10">
-      <div className="w-full">
+    <div className="bg-gradient-to-b from-[#005D99] to-[#005083] h-screen flex flex-col justify-center items-center gap-10 p-12 text-white">
+      <div className="w-full max-w-xl text-center space-y-2">
         <div className="px-3">
           <img
             src={TrisaktiLogo}
-            className="lg:w-36 w-32 -ml-2 flex items-start"
+            className="w-28 mx-auto"
             data-aos="fade-right"
             data-aos-offset="300"
             alt="Logo-Usakti"
           />
         </div>
         <div
-          className="font-bold text-[30px]"
+          className="font-bold text-3xl"
           data-aos="fade-right"
           data-aos-offset="300"
           data-aos-delay="100"
         >
-          Hallo Nabila Azzahra <span className="text-[35px]">🙌</span>
+          <h2>Halo Nabila Azzahra</h2>
         </div>
         <div data-aos="fade-right" data-aos-offset="300" data-aos-delay="100">
-          <p className="text-sm">Lorem ipsum dolor sit amet consectetur adipisicing elit. Cumque
-            similique ipsam eum amet veritatis accusamus possimus itaque
-            repellendus illo sunt exercitationem autem enim quod repellat, eius
-            distinctio ratione ex hic expedita ut nobis dolorem. Nemo rem at
-            molestiae. Facilis repellendus autem odio. Consectetur odit blanditiis
-            laudantium quis soluta quisquam, officiis expedita inventore, iusto
-            magnam explicabo nostrum earum deserunt perspiciatis! Quisquam magnam
-            a ab, enim non modi laborum, est ad illum ut cum magni animi incidunt
-            molestiae? Ratione, ipsa. Optio veniam ut, velit doloremque ab
-            blanditiis, aut dolor iste repellat quis repellendus? Alias corrupti
-            iure vel ut vitae beatae quo veniam?</p>
+          <p className="text-sm">Lorem ipsum dolor sit amet consectetur adipisicing elit. Facilis dolor esse accusantium, doloremque rerum non voluptates saepe distinctio reiciendis corrupti explicabo debitis eaque laborum ab rem odio ipsam sed ea. Accusantium ab illum commodi aliquam quos, fugiat cum doloribus, soluta nulla ipsam omnis blanditiis et voluptatem tempora. Dolorem, aut dolores!</p>
         </div>
-        <div
-          className="flex items-center gap-5 pt-4"
-          data-aos="fade-down"
-          data-aos-duration="1000"
-        >
-          <div>
-            <button className="bg-red-500 text-white w-28 text-sm font-bold rounded-xl flex gap-2 items-center justify-center px-4 py-2">
+        <div className="flex items-center gap-5 pt-4">
+          <div className="mx-auto">
+            <button type="button" data-aos="fade-right" onClick={handleLogout} className="bg-red-500 hover:bg-red-600 transition-all ease-in-out text-white w-28 text-sm font-bold rounded-xl flex gap-2 items-center justify-center px-4 py-2">
               <div className="flex items-center">Keluar</div>
               <div className="pt-1">
                 <i className="fi fi-br-sign-in-alt w-1" />
@@ -114,22 +126,28 @@ const Dashboard = () => {
           </div>
         </div>
       </div>
-      <div className="w-full text-black">
+      <div className="w-full">
         {packages && packages.length > 0 ? (
-          <div className="grid grid-cols-3 w-full gap-2">
+          <div className="flex items-center justify-center flex-wrap">
             {packages.map((pkg, index) => (
               <button
                 type="button"
                 key={index}
-                onClick={() => checkoutPackage(pkg.id)}
-                className="bg-white hover:bg-gray-100 font-medium p-5 rounded-2xl"
+                onClick={() => checkoutPackage(pkg)}
+                className="block w-1/2 md:w-1/6 p-1"
+                data-aos="fade-down"
+                data-aos-delay={index + 1 * 200}
               >
-                {pkg.name || `TPA ${index + 1}`}
+                <span className="block text-gray-800 text-sm bg-white hover:bg-gray-100 font-medium p-5 rounded-2xl">
+                  {pkg.name || `TPA ${index + 1}`}
+                </span>
               </button>
             ))}
           </div>
         ) : (
-          <p>Tidak ada package</p>
+          <p className="text-center text-sm underline"
+            data-aos="fade-down"
+            data-aos-delay={200}>Paket soal belum tersedia, silahkan hubungi panitia PMB.</p>
         )}
       </div>
     </div>
